# api/src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
			"IApplication": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"type": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"appType": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					},
					"platformList": {
						"items": {
							"properties": {
								"appProfile": {
									"properties": {
										"appProfileId": {
											"type": "string"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"appProfileId",
										"name"
									],
									"type": "object"
								},
								"platformName": {
									"type": "string"
								},
								"platformId": {
									"type": "string"
								}
							},
							"required": [
								"appProfile",
								"platformId"
							],
							"type": "object"
						},
						"type": "array"
					},
					"grant_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"type",
					"name",
					"appType",
					"clientId",
					"clientSecret"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IApplicationCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					},
					"appType": {
						"type": "string"
					},
					"redirectUri": {
						"type": "string"
					},
					"grant_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"platformList": {
						"items": {
							"properties": {
								"appProfile": {
									"properties": {
										"appProfileId": {
											"type": "string"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"appProfileId",
										"name"
									],
									"type": "object"
								},
								"platformId": {
									"type": "string"
								},
								"platformName": {
									"type": "string"
								}
							},
							"required": [
								"appProfile",
								"platformId"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"clientId",
					"clientSecret"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IApplicationUpdateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					},
					"appType": {
						"type": "string"
					},
					"redirectUri": {
						"type": "string"
					},
					"platformList": {
						"items": {
							"properties": {
								"appProfile": {
									"properties": {
										"appProfileName": {
											"type": "string"
										},
										"appProfileBosConfigId": {
											"type": "string"
										},
										"appProfileAdminId": {
											"type": "string"
										}
									},
									"required": [
										"appProfileName",
										"appProfileBosConfigId",
										"appProfileAdminId"
									],
									"type": "object"
								},
								"platformName": {
									"type": "string"
								},
								"platformId": {
									"type": "string"
								}
							},
							"required": [
								"appProfile",
								"platformId"
							],
							"type": "object"
						},
						"type": "array"
					},
					"grant_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IToken": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"createdToken": {
						"type": "number",
						"format": "double"
					},
					"expieredToken": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPlatform": {
				"properties": {
					"idPlatformOfAdmin": {
						"type": "string"
					},
					"platformName": {
						"type": "string"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"idPlatformOfAdmin",
					"platformName",
					"platformId"
				],
				"type": "object"
			},
			"IAppProfile": {
				"properties": {
					"appProfileName": {
						"type": "string"
					},
					"appProfileBosConfigId": {
						"type": "string"
					},
					"appProfileAdminId": {
						"type": "string"
					}
				},
				"required": [
					"appProfileName",
					"appProfileBosConfigId",
					"appProfileAdminId"
				],
				"type": "object"
			},
			"IAppPlatformProfile": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IPlatform"
					},
					{
						"properties": {
							"appProfile": {
								"$ref": "#/components/schemas/IAppProfile"
							}
						},
						"required": [
							"appProfile"
						],
						"type": "object"
					}
				]
			},
			"IApplicationToken": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IToken"
					},
					{
						"properties": {
							"platformList": {
								"items": {
									"$ref": "#/components/schemas/IAppPlatformProfile"
								},
								"type": "array"
							},
							"applicationProfileList": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"applicationId": {
								"type": "string"
							}
						},
						"type": "object"
					}
				]
			},
			"IApplicationLoginParams": {
				"properties": {
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					}
				},
				"required": [
					"clientId",
					"clientSecret"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IApplicationLogs": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"actor": {
						"properties": {
							"actorName": {
								"type": "string"
							},
							"actorId": {
								"type": "string"
							}
						},
						"required": [
							"actorName",
							"actorId"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"date",
					"message",
					"actor"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Spinal Organ Auth API",
		"version": "1.0.0",
		"description": "API for managing authentication and authorization in the Spinal Organ system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Spinalcom",
			"email": "developers@spinalcom.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/applications": {
			"post": {
				"operationId": "CreateApplication",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IApplication"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IApplicationCreationParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetApplications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IApplication"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/applications/{applicationId}": {
			"get": {
				"operationId": "GetApplication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IApplication"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read",
							"ownData:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteApplication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										},
										"message": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateApplication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IApplication"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IApplicationUpdateParams"
							}
						}
					}
				}
			}
		},
		"/applications/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IApplicationToken"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IApplicationLoginParams"
							}
						}
					}
				}
			}
		},
		"/applications/{applicationId}/applicationLogs": {
			"get": {
				"operationId": "GetApplicationLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IApplicationLogs"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Applications"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "applicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}